net: lib: http_server: Split HTTP server into multiple files
As the server source file grows it's becoming harder to navigate,
therefore split the source into logical subcomponents (server core,
HTTP1 processing, HTTP2 processing).
