+++ include/zephyr/net/http/server.h
@@ -30,6 +30,8 @@
 #define HTTP_SERVER_MAX_STREAMS        CONFIG_HTTP_SERVER_MAX_STREAMS
 #define HTTP_SERVER_MAX_CONTENT_TYPE_LEN CONFIG_HTTP_SERVER_MAX_CONTENT_TYPE_LENGTH
 
+#define HTTP2_PREFACE "PRI * HTTP/2.0\r\n\r\nSM\r\n\r\n"
+
 enum http_resource_type {
 	HTTP_RESOURCE_TYPE_STATIC,
 	HTTP_RESOURCE_TYPE_DYNAMIC,
