+++ subsys/net/lib/http/CMakeLists.txt
@@ -12,4 +12,9 @@ zephyr_library_sources_ifdef(CONFIG_HTTP_PARSER http_parser.c)
 zephyr_library_sources_ifdef(CONFIG_HTTP_PARSER_URL http_parser_url.c)
 zephyr_library_sources_ifdef(CONFIG_HTTP_CLIENT http_client.c)
 zephyr_library_sources_ifdef(CONFIG_HTTP_HPACK_TABLE http_hpack.c http_huffman.c)
-zephyr_library_sources_ifdef(CONFIG_HTTP_SERVER http_server.c)
+zephyr_library_sources_ifdef(
+  CONFIG_HTTP_SERVER
+  http_server_core.c
+  http_server_http1.c
+  http_server_http2.c
+)
