+++ subsys/net/lib/http/http_server_core.c
@@ -1,0 +1,588 @@
+/*
+ * Copyright (c) 2023, Emna Rekik
+ * Copyright (c) 2024 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <errno.h>
+#include <stdbool.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <strings.h>
+
+#include <zephyr/kernel.h>
+#include <zephyr/logging/log.h>
+#include <zephyr/net/http/service.h>
+#include <zephyr/net/net_ip.h>
+#include <zephyr/net/socket.h>
+#include <zephyr/net/tls_credentials.h>
+#include <zephyr/posix/sys/eventfd.h>
+
+LOG_MODULE_REGISTER(net_http_server, CONFIG_NET_HTTP_SERVER_LOG_LEVEL);
+
+#include "../../ip/net_private.h"
+#include "headers/server_functions.h"
+
+#define INVALID_SOCK -1
+
+int http_server_init(struct http_server_ctx *ctx)
+{
+	int proto;
+	int failed = 0, count = 0;
+	int svc_count;
+	socklen_t len;
+	int fd, af, i;
+	struct sockaddr_storage addr_storage;
+	const union {
+		struct sockaddr *addr;
+		struct sockaddr_in *addr4;
+		struct sockaddr_in6 *addr6;
+	} addr = {
+		.addr = (struct sockaddr *)&addr_storage
+	};
+
+	HTTP_SERVICE_COUNT(&svc_count);
+
+	/* Initialize fds */
+	memset(ctx->fds, 0, sizeof(ctx->fds));
+	memset(ctx->clients, 0, sizeof(ctx->clients));
+
+	for (i = 0; i < ARRAY_SIZE(ctx->fds); i++) {
+		ctx->fds[i].fd = INVALID_SOCK;
+	}
+
+	/* Create an eventfd that can be used to trigger events during polling */
+	fd = eventfd(0, 0);
+	if (fd < 0) {
+		fd = -errno;
+		LOG_ERR("eventfd failed (%d)", fd);
+		return fd;
+	}
+
+	ctx->fds[count].fd = fd;
+	ctx->fds[count].events = POLLIN;
+	count++;
+
+	HTTP_SERVICE_FOREACH(svc) {
+		/* set the default address (in6addr_any / INADDR_ANY are all 0) */
+		memset(&addr_storage, 0, sizeof(struct sockaddr_storage));
+
+		/* Set up the server address struct according to address family */
+		if (IS_ENABLED(CONFIG_NET_IPV6) &&
+		    zsock_inet_pton(AF_INET6, svc->host, &addr.addr6->sin6_addr) == 1) {
+			af = AF_INET6;
+			len = sizeof(*addr.addr6);
+
+			addr.addr6->sin6_family = AF_INET6;
+			addr.addr6->sin6_port = htons(*svc->port);
+		} else if (IS_ENABLED(CONFIG_NET_IPV4) &&
+			   zsock_inet_pton(AF_INET, svc->host, &addr.addr4->sin_addr) == 1) {
+			af = AF_INET;
+			len = sizeof(*addr.addr4);
+
+			addr.addr4->sin_family = AF_INET;
+			addr.addr4->sin_port = htons(*svc->port);
+		} else if (IS_ENABLED(CONFIG_NET_IPV6)) {
+			/* prefer IPv6 if both IPv6 and IPv4 are supported */
+			af = AF_INET6;
+			len = sizeof(*addr.addr6);
+
+			addr.addr6->sin6_family = AF_INET6;
+			addr.addr6->sin6_port = htons(*svc->port);
+		} else if (IS_ENABLED(CONFIG_NET_IPV4)) {
+			af = AF_INET;
+			len = sizeof(*addr.addr4);
+
+			addr.addr4->sin_family = AF_INET;
+			addr.addr4->sin_port = htons(*svc->port);
+		} else {
+			LOG_ERR("Neither IPv4 nor IPv6 is enabled");
+			failed++;
+			break;
+		}
+
+		/* Create a socket */
+		if (COND_CODE_1(CONFIG_NET_SOCKETS_SOCKOPT_TLS,
+				(svc->sec_tag_list != NULL),
+				(0))) {
+			proto = IPPROTO_TLS_1_2;
+		} else {
+			proto = IPPROTO_TCP;
+		}
+
+		fd = zsock_socket(af, SOCK_STREAM, proto);
+		if (fd < 0) {
+			LOG_ERR("socket: %d", errno);
+			failed++;
+			continue;
+		}
+
+#if defined(CONFIG_NET_SOCKETS_SOCKOPT_TLS)
+		if (svc->sec_tag_list != NULL) {
+			if (zsock_setsockopt(fd, SOL_TLS, TLS_SEC_TAG_LIST,
+					     svc->sec_tag_list,
+					     svc->sec_tag_list_size) < 0) {
+				LOG_ERR("setsockopt: %d", errno);
+				zsock_close(fd);
+				continue;
+			}
+
+			if (zsock_setsockopt(fd, SOL_TLS, TLS_HOSTNAME, "localhost",
+					     sizeof("localhost")) < 0) {
+				LOG_ERR("setsockopt: %d", errno);
+				zsock_close(fd);
+				continue;
+			}
+		}
+#endif
+
+		if (zsock_setsockopt(fd, SOL_SOCKET, SO_REUSEADDR, &(int){1},
+				     sizeof(int)) < 0) {
+			LOG_ERR("setsockopt: %d", errno);
+			zsock_close(fd);
+			continue;
+		}
+
+		if (zsock_bind(fd, addr.addr, len) < 0) {
+			LOG_ERR("bind: %d", errno);
+			failed++;
+			zsock_close(fd);
+			continue;
+		}
+
+		if (*svc->port == 0) {
+			/* ephemeral port - read back the port number */
+			len = sizeof(addr_storage);
+			if (zsock_getsockname(fd, addr.addr, &len) < 0) {
+				LOG_ERR("getsockname: %d", errno);
+				zsock_close(fd);
+				continue;
+			}
+
+			*svc->port = ntohs(addr.addr4->sin_port);
+		}
+
+		if (zsock_listen(fd, HTTP_SERVER_MAX_CLIENTS) < 0) {
+			LOG_ERR("listen: %d", errno);
+			failed++;
+			zsock_close(fd);
+			continue;
+		}
+
+		LOG_DBG("Initialized HTTP Service %s:%u", svc->host, *svc->port);
+
+		ctx->fds[count].fd = fd;
+		ctx->fds[count].events = POLLIN;
+		count++;
+	}
+
+	if (failed >= svc_count) {
+		LOG_ERR("All services failed (%d)", failed);
+		return -ESRCH;
+	}
+
+	ctx->listen_fds = count;
+	ctx->num_clients = 0;
+
+	return 0;
+}
+
+static int accept_new_client(int server_fd)
+{
+	int new_socket;
+	socklen_t addrlen;
+	struct sockaddr_storage sa;
+
+	memset(&sa, 0, sizeof(sa));
+	addrlen = sizeof(sa);
+
+	new_socket = zsock_accept(server_fd, (struct sockaddr *)&sa, &addrlen);
+	if (new_socket < 0) {
+		new_socket = -errno;
+		LOG_DBG("[%d] accept failed (%d)", server_fd, new_socket);
+		return new_socket;
+	}
+
+	LOG_DBG("New client from %s:%d",
+		net_sprint_addr(sa.ss_family, &net_sin((struct sockaddr *)&sa)->sin_addr),
+		ntohs(net_sin((struct sockaddr *)&sa)->sin_port));
+
+	return new_socket;
+}
+
+static int close_all_sockets(struct http_server_ctx *ctx)
+{
+	close(ctx->fds[0].fd); /* close eventfd */
+	ctx->fds[0].fd = -1;
+
+	for (int i = 1; i < ARRAY_SIZE(ctx->fds); i++) {
+		if (ctx->fds[i].fd < 0) {
+			continue;
+		}
+
+		zsock_close(ctx->fds[i].fd);
+		ctx->fds[i].fd = -1;
+	}
+
+	return 0;
+}
+
+static void init_client_ctx(struct http_client_ctx *client, int new_socket)
+{
+	client->fd = new_socket;
+	client->data_len = 0;
+	client->server_state = HTTP_SERVER_PREFACE_STATE;
+	client->has_upgrade_header = false;
+	client->preface_sent = false;
+
+	memset(client->buffer, 0, sizeof(client->buffer));
+	memset(client->url_buffer, 0, sizeof(client->url_buffer));
+
+	ARRAY_FOR_EACH(client->streams, i) {
+		client->streams[i].stream_state = HTTP_SERVER_STREAM_IDLE;
+		client->streams[i].stream_id = 0;
+	}
+}
+
+static void close_client_connection(struct http_server_ctx *server,
+				    struct http_client_ctx *client)
+{
+	int i;
+
+	__ASSERT_NO_MSG(IS_ARRAY_ELEMENT(server->clients, client));
+
+	zsock_close(client->fd);
+
+	server->num_clients--;
+
+	for (i = server->listen_fds; i < ARRAY_SIZE(server->fds); i++) {
+		if (server->fds[i].fd == client->fd) {
+			server->fds[i].fd = INVALID_SOCK;
+			break;
+		}
+	}
+
+	memset(client, 0, sizeof(struct http_client_ctx));
+}
+
+static int handle_http_preface(struct http_server_ctx *server,
+			       struct http_client_ctx *client)
+{
+	LOG_DBG("HTTP_SERVER_PREFACE_STATE.");
+
+	if (client->data_len < sizeof(HTTP2_PREFACE) - 1) {
+		/* We don't have full preface yet, get more data. */
+		return -EAGAIN;
+	}
+
+	if (strncmp(client->cursor, HTTP2_PREFACE, sizeof(HTTP2_PREFACE) - 1) != 0) {
+		return enter_http1_request(client);
+	}
+
+	return enter_http2_request(client);
+}
+
+static int handle_http_done(struct http_server_ctx *server,
+			    struct http_client_ctx *client)
+{
+	LOG_DBG("HTTP_SERVER_DONE_STATE");
+
+	close_client_connection(server, client);
+
+	return -EAGAIN;
+}
+
+int enter_http_done_state(struct http_server_ctx *server,
+			  struct http_client_ctx *client)
+{
+	client->server_state = HTTP_SERVER_DONE_STATE;
+
+	close_client_connection(server, client);
+
+	return 0;
+}
+
+static int handle_http_request(struct http_server_ctx *server, struct http_client_ctx *client)
+{
+	int ret = -EINVAL;
+
+	client->cursor = client->buffer;
+
+	do {
+		switch (client->server_state) {
+		case HTTP_SERVER_FRAME_DATA_STATE:
+			ret = handle_http_frame_data(client);
+			break;
+		case HTTP_SERVER_PREFACE_STATE:
+			ret = handle_http_preface(server, client);
+			break;
+		case HTTP_SERVER_REQUEST_STATE:
+			ret = handle_http1_request(server, client);
+			break;
+		case HTTP_SERVER_FRAME_HEADER_STATE:
+			ret = handle_http_frame_header(server, client);
+			break;
+		case HTTP_SERVER_FRAME_HEADERS_STATE:
+			ret = handle_http_frame_headers(client);
+			break;
+		case HTTP_SERVER_FRAME_CONTINUATION_STATE:
+			ret = handle_http_frame_continuation(client);
+			break;
+		case HTTP_SERVER_FRAME_SETTINGS_STATE:
+			ret = handle_http_frame_settings(client);
+			break;
+		case HTTP_SERVER_FRAME_WINDOW_UPDATE_STATE:
+			ret = handle_http_frame_window_update(client);
+			break;
+		case HTTP_SERVER_FRAME_RST_STREAM_STATE:
+			ret = handle_http_frame_rst_frame(server, client);
+			break;
+		case HTTP_SERVER_FRAME_GOAWAY_STATE:
+			ret = handle_http_frame_goaway(server, client);
+			break;
+		case HTTP_SERVER_FRAME_PRIORITY_STATE:
+			ret = handle_http_frame_priority(client);
+			break;
+		case HTTP_SERVER_DONE_STATE:
+			ret = handle_http_done(server, client);
+			break;
+		default:
+			ret = handle_http_done(server, client);
+			break;
+		}
+	} while (ret >= 0 && client->data_len > 0);
+
+	if (ret < 0 && ret != -EAGAIN) {
+		return ret;
+	}
+
+	if (client->data_len > 0) {
+		/* Move any remaining data in the buffer. */
+		memmove(client->buffer, client->cursor, client->data_len);
+	}
+
+	return 0;
+}
+
+int http_server_start(struct http_server_ctx *ctx)
+{
+	struct http_client_ctx *client;
+	eventfd_t value;
+	bool found_slot;
+	int new_socket;
+	int ret, i, j;
+	int sock_error;
+	socklen_t optlen = sizeof(int);
+
+	value = 0;
+
+	while (1) {
+		ret = zsock_poll(ctx->fds, ctx->listen_fds + ctx->num_clients, -1);
+		if (ret < 0) {
+			ret = -errno;
+			LOG_DBG("poll failed (%d)", ret);
+			return ret;
+		}
+
+		if (ret == 0) {
+			/* should not happen because timeout is -1 */
+			break;
+		}
+
+		if (ret == 1 && ctx->fds[0].revents) {
+			eventfd_read(ctx->fds[0].fd, &value);
+			LOG_DBG("Received stop event. exiting ..");
+			goto closing;
+		}
+
+		for (i = 1; i < ARRAY_SIZE(ctx->fds); i++) {
+			if (ctx->fds[i].fd < 0) {
+				continue;
+			}
+
+			if (ctx->fds[i].revents & POLLHUP) {
+				if (i >= ctx->listen_fds) {
+					LOG_DBG("Client #%d has disconnected",
+						i - ctx->listen_fds);
+
+					client = &ctx->clients[i - ctx->listen_fds];
+					close_client_connection(ctx, client);
+				}
+
+				continue;
+			}
+
+			if (ctx->fds[i].revents & POLLERR) {
+				(void)zsock_getsockopt(ctx->fds[i].fd, SOL_SOCKET,
+						       SO_ERROR, &sock_error, &optlen);
+				LOG_DBG("Error on fd %d %d", ctx->fds[i].fd, sock_error);
+
+				if (i >= ctx->listen_fds) {
+					client = &ctx->clients[i - ctx->listen_fds];
+					close_client_connection(ctx, client);
+					continue;
+				}
+
+				/* Listening socket error, abort. */
+				LOG_ERR("Listening socket error, aborting.");
+				return -sock_error;
+
+			}
+
+			if (!(ctx->fds[i].revents & POLLIN)) {
+				continue;
+			}
+
+			/* First check if we have something to accept */
+			if (i < ctx->listen_fds) {
+				new_socket = accept_new_client(ctx->fds[i].fd);
+				if (new_socket < 0) {
+					ret = -errno;
+					LOG_DBG("accept: %d", ret);
+					continue;
+				}
+
+				found_slot = false;
+
+				for (j = ctx->listen_fds; j < ARRAY_SIZE(ctx->fds); j++) {
+					if (ctx->fds[j].fd != INVALID_SOCK) {
+						continue;
+					}
+
+					ctx->fds[j].fd = new_socket;
+					ctx->fds[j].events = POLLIN;
+					ctx->fds[j].revents = 0;
+
+					ctx->num_clients++;
+
+					LOG_DBG("Init client #%d", j - ctx->listen_fds);
+
+					init_client_ctx(&ctx->clients[j - ctx->listen_fds],
+							new_socket);
+					found_slot = true;
+					break;
+				}
+
+				if (!found_slot) {
+					LOG_DBG("No free slot found.");
+					zsock_close(new_socket);
+				}
+
+				continue;
+			}
+
+			/* Client sock */
+			client = &ctx->clients[i - ctx->listen_fds];
+
+			ret = zsock_recv(client->fd, client->buffer + client->data_len,
+					 sizeof(client->buffer) - client->data_len, 0);
+			if (ret <= 0) {
+				if (ret == 0) {
+					LOG_DBG("Connection closed by peer for client #%d",
+						i - ctx->listen_fds);
+				} else {
+					ret = -errno;
+					LOG_DBG("ERROR reading from socket (%d)", ret);
+				}
+
+				close_client_connection(ctx, client);
+				continue;
+			}
+
+			client->data_len += ret;
+
+			ret = handle_http_request(ctx, client);
+			if (ret < 0 && ret != -EAGAIN) {
+				if (ret == -ENOTCONN) {
+					LOG_DBG("Client closed connection while handling request");
+				} else {
+					LOG_ERR("HTTP request handling error (%d)", ret);
+				}
+				close_client_connection(ctx, client);
+			} else if (client->data_len == sizeof(client->buffer)) {
+				/* If the RX buffer is still full after parsing,
+				 * it means we won't be able to handle this request
+				 * with the current buffer size.
+				 */
+				LOG_ERR("RX buffer too small to handle request");
+				close_client_connection(ctx, client);
+			}
+		}
+	}
+
+	return 0;
+
+closing:
+	/* Close all client connections and the server socket */
+	return close_all_sockets(ctx);
+}
+
+/* Close all client connections and the server socket + any other things needed */
+int http_server_cleanup(struct http_server_ctx *ctx)
+{
+	if (ctx == NULL) {
+		return -EINVAL;
+	}
+
+	return close_all_sockets(ctx);
+}
+
+int http_server_stop(struct http_server_ctx *ctx)
+{
+	eventfd_write(ctx->fds[0].fd, 1);
+
+	return 0;
+}
+
+/* Compare two strings where the terminator is either "\0" or "?" */
+static int compare_strings(const char *s1, const char *s2)
+{
+	while ((*s1 && *s2) && (*s1 == *s2) && (*s1 != '?')) {
+		s1++;
+		s2++;
+	}
+
+	/* Check if both strings have reached their terminators or '?' */
+	if ((*s1 == '\0' || *s1 == '?') && (*s2 == '\0' || *s2 == '?')) {
+		return 0; /* Strings are equal */
+	}
+
+	return 1; /* Strings are not equal */
+}
+
+struct http_resource_detail *get_resource_detail(const char *path,
+						 int *path_len)
+{
+	HTTP_SERVICE_FOREACH(service) {
+		HTTP_SERVICE_FOREACH_RESOURCE(service, resource) {
+			if (compare_strings(path, resource->resource) == 0) {
+				NET_DBG("Got match for %s", resource->resource);
+
+				*path_len = strlen(resource->resource);
+				return resource->detail;
+			}
+		}
+	}
+
+	NET_DBG("No match for %s", path);
+
+	return NULL;
+}
+
+int http_server_sendall(int sock, const void *buf, size_t len)
+{
+	while (len) {
+		ssize_t out_len = zsock_send(sock, buf, len, 0);
+
+		if (out_len < 0) {
+			return -errno;
+		}
+
+		buf = (const char *)buf + out_len;
+		len -= out_len;
+	}
+
+	return 0;
+}
